achieved more numerically stable marching

+++ src/shaders/mpr_viewer/surface/chunks/uniforms/uniforms_volume.glsl
@@ -3,7 +3,7 @@
 
 struct Volume 
 {
-    vec3  dimensions;    
+    ivec3 dimensions;    
     vec3  spacing;           
     vec3  size;         
     vec3  inv_dimensions;
